(fix): middleware와 comment로직 간 형변환 버그 수정

+++ app/api/comment/resource/request.go
@@ -2,30 +2,30 @@ package resource
 
 type FindAllCommentRequest struct {
 	// 라우터에 포함될 EventId Param
-	EventId int `json:"event_id" param:"id"`
+	EventId uint64 `json:"event_id" param:"id"`
 }
 
 type CreateCommentRequest struct {
 	// 라우터에 포함될 EventId Param
-	EventId int `json:"event_id" param:"id"`
+	EventId uint64 `json:"event_id" param:"id"`
 	// 등록할 UserId와 Content body
-	UserId  int    `json:"user_id"`
+	UserId  uint64 `json:"user_id"`
 	Content string `json:"content"`
 }
 
 type UpdateCommentRequest struct {
-	Id int `json:"id"`
+	Id uint64 `json:"id"`
 	// 라우터에 포함될 Id Param
-	EventId int `json:"event_id" param:"id"`
+	EventId uint64 `json:"event_id" param:"id"`
 	// 수정할 UserId와 Content body
-	UserId  int    `json:"user_id"`
+	UserId  uint64 `json:"user_id"`
 	Content string `json:"content"`
 }
 
 type DeleteCommentRequest struct {
-	Id int `json:"id"`
+	Id uint64 `json:"id"`
 	// 라우터에 포함될 Id Param
-	EventId int `json:"event_id" param:"id"`
+	EventId uint64 `json:"event_id" param:"id"`
 	// 삭제할 UserId body
-	UserId int `json:"user_id"`
+	UserId uint64 `json:"user_id"`
 }
